refactor: 1단계 피드백 반영
- 줄 간격 통해 함수 속 기능 구분
- dimmer 클릭 시 모달창 닫히도록 구현
- 무거운 index.js 분리
